replace FSA with Array2 in Determinize (#63)
